Change comment and programming style

+++ linux_prereqs_installer.sh
@@ -1,6 +1,4 @@
 #!/bin/bash
-
-################################################################################
 #
 # Installs all of the packages and dependencies required for NadekoBot to run
 # on Linux based distributions.
@@ -9,16 +7,61 @@
 # 'linuxAIO.sh', 'installer_prep.sh', and 'nadeko_master_installer.sh'.
 #
 ################################################################################
-#
-# [ Functions ]
-#
-###
-    dot_net_install() {
+#### [ Functions ]
+
+
+dot_net_install() {
+    echo "Installing .NET Core..."
+    # Microsoft package signing key
+    curl https://packages.microsoft.com/config/"$1"/"$2"/packages-microsoft-prod.deb \
+        -o packages-microsoft-prod.deb
+    sudo dpkg -i packages-microsoft-prod.deb && sudo rm -f packages-microsoft-prod.deb
+    # Install the SDK
+    sudo apt-get update
+    sudo apt-get install -y apt-transport-https &&
+    sudo apt-get update &&
+    sudo apt-get install -y dotnet-sdk-3.1
+
+    echo "Installing prerequisites..."
+    sudo apt-get install libopus0 opus-tools libopus-dev libsodium-dev ffmpeg \
+        redis-server git python python3 jq wget -y
+
+    sudo curl -s -L https://yt-dl.org/downloads/latest/youtube-dl -o /usr/local/bin/youtube-dl
+    sudo chmod a+rx /usr/local/bin/youtube-dl
+}
+
+
+#### End of [ Functions ]
+################################################################################
+#### [ Main ]
+
+
+read -p "We will now install NadekoBot's prerequisites. Press [Enter] to continue."
+
+if [[ $distro = "ubuntu" ]]; then
+    if [[ $ver = "16.04" ]]; then
+        dot_net_install "ubuntu" "16.04"
+    elif [[ $ver = "18.04" ]]; then
+        dot_net_install "ubuntu" "18.04"
+    elif [[ $ver = "20.04" ]]; then
+        dot_net_install "ubuntu" "20.04"
+    else
+        echo "${red}The installer does not support the automatic" \
+            "installation and setup of NadekoBot's prerequisites for your" \
+            "OS: $distro $ver $arch${nc}"
+        read -p "Press [Enter] to return to the installer menu"
+    fi
+elif [[ $distro = "debian" ]]; then
+    if [[ $sver = "9" ]]; then
         echo "Installing .NET Core..."
         # Microsoft package signing key
-        curl https://packages.microsoft.com/config/"$1"/"$2"/packages-microsoft-prod.deb \
-            -o packages-microsoft-prod.deb
-        sudo dpkg -i packages-microsoft-prod.deb && sudo rm -f packages-microsoft-prod.deb
+        wget -O - https://packages.microsoft.com/keys/microsoft.asc | gpg \
+            --dearmor >microsoft.asc.gpg
+        sudo mv microsoft.asc.gpg /etc/apt/trusted.gpg.d/
+        curl -s https://packages.microsoft.com/config/debian/9/prod.list -o prod.list
+        sudo mv prod.list /etc/apt/sources.list.d/microsoft-prod.list
+        sudo chown root:root /etc/apt/trusted.gpg.d/microsoft.asc.gpg
+        sudo chown root:root /etc/apt/sources.list.d/microsoft-prod.list
         # Install the SDK
         sudo apt-get update
         sudo apt-get install -y apt-transport-https &&
@@ -26,87 +69,37 @@
         sudo apt-get install -y dotnet-sdk-3.1
 
         echo "Installing prerequisites..."
-        sudo apt-get install libopus0 opus-tools libopus-dev libsodium-dev ffmpeg \
-            redis-server git python python3 jq wget -y
-
-        sudo curl -s -L https://yt-dl.org/downloads/latest/youtube-dl -o /usr/local/bin/youtube-dl
+        sudo apt-get install libopus0 opus-tools libopus-dev libsodium-dev \
+            ffmpeg redis-server git jq python python3 -y
+        sudo curl -s -L https://yt-dl.org/downloads/latest/youtube-dl -o \
+            /usr/local/bin/youtube-dl
         sudo chmod a+rx /usr/local/bin/youtube-dl
-    }
-###
-#
-# End of [ Functions ]
-################################################################################
-
-
-################################################################################
-#
-# [ Main ]
-#
-###
-    read -p "We will now install NadekoBot's prerequisites. Press [Enter] to continue."
-
-    if [[ $distro = "ubuntu" ]]; then
-        if [[ $ver = "16.04" ]]; then
-            dot_net_install "ubuntu" "16.04"
-        elif [[ $ver = "18.04" ]]; then
-            dot_net_install "ubuntu" "18.04"
-        elif [[ $ver = "20.04" ]]; then
-            dot_net_install "ubuntu" "20.04"
-        else
-            echo "${red}The installer does not support the automatic" \
-                "installation and setup of NadekoBot's prerequisites for your" \
-                "OS: $distro $ver $arch${nc}"
-            read -p "Press [Enter] to return to the installer menu"
-        fi
-    elif [[ $distro = "debian" ]]; then
-        if [[ $sver = "9" ]]; then
-            echo "Installing .NET Core..."
-            # Microsoft package signing key
-            wget -O - https://packages.microsoft.com/keys/microsoft.asc | gpg \
-                --dearmor >microsoft.asc.gpg
-            sudo mv microsoft.asc.gpg /etc/apt/trusted.gpg.d/
-            curl -s https://packages.microsoft.com/config/debian/9/prod.list -o prod.list
-            sudo mv prod.list /etc/apt/sources.list.d/microsoft-prod.list
-            sudo chown root:root /etc/apt/trusted.gpg.d/microsoft.asc.gpg
-            sudo chown root:root /etc/apt/sources.list.d/microsoft-prod.list
-            # Install the SDK
-            sudo apt-get update
-            sudo apt-get install -y apt-transport-https &&
-            sudo apt-get update &&
-            sudo apt-get install -y dotnet-sdk-3.1
-
-            echo "Installing prerequisites..."
-            sudo apt-get install libopus0 opus-tools libopus-dev libsodium-dev \
-                ffmpeg redis-server git jq python python3 -y
-            sudo curl -s -L https://yt-dl.org/downloads/latest/youtube-dl -o \
-                /usr/local/bin/youtube-dl
-            sudo chmod a+rx /usr/local/bin/youtube-dl
-        elif [[ $sver = "10" ]]; then
-            dot_net_install "debian" "10"
-        else
-            echo "${red}The installer does not support the automatic" \
-                "installation and setup of NadekoBot's prerequisites for your" \
-                "OS: $distro $ver $arch${nc}"
-            read -p "Press [Enter] to return to the installer menu"
-        fi
-    elif [[ $distro = "linuxmint" ]]; then
-        if [[ $sver = "18" ]]; then
-            dot_net_install "ubuntu" "16.04"
-        elif [[ $sver = "19" ]]; then
-            dot_net_install "ubuntu" "18.04"
-        elif [[ $sver = "20" ]]; then
-            dot_net_install "ubuntu" "20.04"
-        else
-            echo "${red}The installer does not support the automatic" \
-                "installation and setup of NadekoBot's prerequisites for your" \
-                "OS: $distro $ver $arch${nc}"
-            read -p "Press [Enter] to return to the installer menu"
-        fi
+    elif [[ $sver = "10" ]]; then
+        dot_net_install "debian" "10"
+    else
+        echo "${red}The installer does not support the automatic" \
+            "installation and setup of NadekoBot's prerequisites for your" \
+            "OS: $distro $ver $arch${nc}"
+        read -p "Press [Enter] to return to the installer menu"
     fi
+elif [[ $distro = "linuxmint" ]]; then
+    if [[ $sver = "18" ]]; then
+        dot_net_install "ubuntu" "16.04"
+    elif [[ $sver = "19" ]]; then
+        dot_net_install "ubuntu" "18.04"
+    elif [[ $sver = "20" ]]; then
+        dot_net_install "ubuntu" "20.04"
+    else
+        echo "${red}The installer does not support the automatic" \
+            "installation and setup of NadekoBot's prerequisites for your" \
+            "OS: $distro $ver $arch${nc}"
+        read -p "Press [Enter] to return to the installer menu"
+    fi
+fi
 
-    echo -e "\n${green}Finished installing prerequisites${nc}"
-    read -p "Press [Enter] to return to the installer menu"
-###
-#
-# End of [ Main ]
+echo -e "\n${green}Finished installing prerequisites${nc}"
+read -p "Press [Enter] to return to the installer menu"
+
+
+#### End of [ Main ]
 ################################################################################
